+++ read_nonblock_windows.go
@@ -1,0 +1,8 @@
+// Copyright 2026 Mariot Chauvin
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build windows
+
+package main
+
+const oReadNonblock = 0